Noticed through inspection of the Flexus .so binary that there were still Simics API calls being made, despite my best efforts to replace everything with QEMU API calls.
The problem seems to be in components/MagicBreakQEMU/breakpoint_tracker.cpp. The issue
with QEMU_insert_callback is that it is a C function, and when the callback is executed,
it cannot execute a function that is the method of a class. So, to fix this, I have
modified the API in qemu/libqemuflex/api.c to include a void* pointer to an object.
If no object is specified (pointer is NULL), then it is ignored. Otherwise, it is
passed as the first argument to the callback function. This way, you can call object
methods using C wrappers defined in the same file, an example being:

class QemuTracer {
	cycles_t trace_mem_hier_operate(conf_object_t *space, generic_transaction_t *trans);
}

extern "C" {

void QemuTracerTraceMemHier(void *obj, conf_object_t *space, generic_transaction_t *trans)
{
	static_cast<QemuTracerImpl*>(obj)->trace_mem_hier(space, trans);
}

}

This way we don't have to mess around with attempting to call class methods from C,
which is impossible. The only problem is that I haven't quite figured out how to keep
the compiler from complaining about this. This was the solution to the problem pretty
much everywhere I looked; I think there's something about casting classes and other pointers
in C++ that I'm overlooking / don't understand.

You can see the changes I made to the QEMU API (and in the other calls to
QEMU_insert_callback in the Flexus code) to get this change to work in the case where
objects are used, and in the case where they aren't.

Good luck. Have fun.