Fix trace mode pagewalk memory leak
When performing page walks in trace mode the PageWalk::preWalk
function is called. This function pushes processed translation
requests to the page walker memory translation
queue (pageWalk.hpp:52). This queue is normally consumed by the MMU
drive function. Since the MMU drive isn't triggered in trace mode the
queue grows indefinitely.

This patch adds a flag to the Translation request container
indicating if a request was made in trace mode. If that's the case,
the request isn't pushed to the memory translation queue.