refactor: rename QEMU <-> Flexus api
This is the single most important commit of the last 4 years.
This update the api and function that are called inside the 'core'
main loop, as well as the components. It was first authored for RISC-V.
This change has to be replicated on the Qemu side.

The code compile for TRACE and TIMING but was not tested.