ci: add build workflow
This test the build system for both
Keen- and KnottyKraken.

ci: fix build call

Use -pr instead of -pr:a.
This will use the profile as the host profile(target machine)
instead of using this for both the (the one who) build and the host.
Which SHOULD trigger a mismatch error when the build compiler
cannot support the host compiler.

ci: use default cmake