fix: C&S operand mismatch (#53)
Found the issue in the internal dependency system for one of the
operands of CAS (3rd operand). It was setting the internal dep arg to 0
which made one operand mark another one as ready (the first one that was
waiting on read register finishing in an extra cycle). Just forcing it
to be the correct arg after the fact fixed the issue, investigating more
on why arg is being set wrong

Co-authored-by: Bryan Perdrizat <bryan.perdrizat@epfl.ch>