fix: LDP/STP forwarding #90
*Description*
LDP and STP instructions load or store to 2 memory locations. It is
normally implemented as two uOps in hardware, but it is a single uOp in
flexus. This causes some unintended bugs where store-load forwarding
breaks etc.

*Steps to Reproduce*
Run data caching image with 2 coresAfter some time, there will be a
sequence of instructions as follows:str to address x + 8ldp to address x

*Expected Behavior*
Load pair should get a part of its value from str because the second
load value of the pair address matches the str address.

*Actual Behavior*
It does not forward because the addresses x and x+8 are not the same.